+++ Interactive-Evolution-SFX/src/Sound.h
@@ -28,7 +28,7 @@ namespace IESFX
 		void load_buffer(const sf::Int16* samples, sf::Uint64 sample_count, unsigned int channel_count = 1, unsigned int sample_rate = SAMPLE_RATE);
 
 		sf::SoundSource::Status status() const;
-		void set_volume(double volume);
+		void set_volume(float volume);
 
 		const sf::Int16* buffer_samples() const { return _buffer.getSamples(); }
 		sf::Uint64 buffer_count() const { return _buffer.getSampleCount(); }
